[REFACTOR(LottoResult)] 패키지 수정
